+++ config/splat.us.ric.yaml
@@ -47,6 +47,7 @@ segments:
       - [0x199AC, .data, 2A060]
       - [0x19D94, .data, 2C4C4]
       - [0x19EE0, .data, 319C4]
+      - [0x19EFC, .data, e_giant_spinning_cross]
       - [0x1A444, .rodata, 1AC60] #UpdateEntityRichter
       - [0x1A5E0, .rodata, 1AC60]
       - [0x1A5F0, .rodata, 1CB04]
@@ -66,8 +67,8 @@ segments:
       - [0x1AB24, .rodata, 2C4C4] #EntitySubwpnAgunea
       - [0x1AB38, .rodata, 319C4]
       - [0x1AB50, .rodata, 319C4]
-      - [0x1AB68, .rodata, 319C4]
-      - [0x1AC50, .rodata, 319C4]
+      - [0x1AB68, .rodata, e_giant_spinning_cross]
+      - [0x1AC50, .rodata, e_giant_spinning_cross]
       - [0x1AC60, c, 1AC60]
       - [0x1CB04, c, 1CB04]
       - [0x1FCD0, c, 1FCD0]
@@ -79,5 +80,6 @@ segments:
       - [0x2A060, c, 2A060]
       - [0x2C4C4, c, 2C4C4]
       - [0x319C4, c, 319C4]
+      - [0x3328C, c, e_giant_spinning_cross]
       - [0x37B64, .bss, bss]
   - [0x39A58]
